app/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart
python-dotenv
loguru
pydantic-settings
